requirements.txt
django >=1.9, <2
azure-storage-blob